spm-tg-bot/requirements.txt
pyTelegramBotAPI>=4.0.0  # Убедитесь, что версия актуальна
requests>=2.26.0          # Убедитесь, что версия актуальна
python-telegram-bot>=21.6
python-dotenv>=1.0.1
telebot>=0.0.5
setuptools>=75.3.0
pika~=1.3.2